pyproject.toml
[project]
name = "image-background-removal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "backgroundremover",
    "numpy>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pytest>=8.4.0",
    "rembg",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]
